Add documentation

+++ SDKBuildScripts/Update-PlayFabSdk.ps1
@@ -1,6 +1,55 @@
+<#
+.SYNOPSIS
+Generate the PlayFab SDK for a platform or language
+
+.DESCRIPTION
+Executes the PlayFab SDK generator to create a new copy of the SDK.  This can be targetted
+against a specific PlayFab vertical, or against APISpec documents generated locally or
+checked into the git repository.
+
+For some platforms, in order use the SDK after generating it you must build it and you need
+some project scaffolding in order to do so.  This scaffolding comes from the SDK specific
+PlayFab repositories.  This script will clone those repositories if necessary and generate
+the new SDK overtop of the existing one.  This is useful to determine what changes have
+occurred in the SDK.
+
+.PARAMETER SdkNames
+The names of one or more of the supported SDKs to generate an SDK for.
+
+.PARAMETER ApiSpecPath
+The path to a local set of API specifications which will be used to generate the SDK.
+
+.PARAMETER ApiSpecPfUrl
+The full URL to the API Spec endpoints which will be used to download the API specification
+files used to generate the SDK.
+
+.PARAMETER ApiSpecPfUrlBranch
+The branch/vertical of the API Spec endpoints which will be used to download the API
+specification files used to generate the SDK. This is a short-form for ApiSpecPfUrl and is
+equivalent to using the URL "https://{ApiSpecPfUrlBranch}.playfabapi.com/apispec".
+
+.PARAMETER ApiSpecGitUrl
+Indicates that the API specifications checked into Git should be used to generate the SDK.
+This is the default behavior and will be used if no other ApiSpec parameter is provided.
+
+.PARAMETER OutputPath
+The location that the generated SDKs will be written to.  This is a base folder path.  A new
+folder under this path will be generated for each provided SDK platform.  The default path
+points to a sibling directory to the SDKGeneratory repository called "sdks".
+
+.PARAMETER KeepSource
+Indicites whether or not to remove all generated source files from the destination location.
+
+.PARAMETER NonNullable
+Indicates that we should generate the SDK for a platform that does not support nullable types.
+
+.PARAMETER Beta
+Indicates whether or not to include any APIs tagged with as beta.
+
+#>
 [CmdletBinding(SupportsShouldProcess = $true, DefaultParameterSetName = "ApiSpecGitUrl")]
 param(
-    [Parameter(Position = 1)]
+    [Parameter(Position = 1, ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true, Mandatory)]
     [ValidateSet(
         "ActionScriptSDK",
         "Cocos2d-xSDK",
@@ -19,14 +68,15 @@ param(
         "WindowsSDK",
         "XPlatCppSDK")]
     [string[]]$SdkNames,
-    [Parameter(ParameterSetName="ApiSpecPath")]
+    [Parameter(ParameterSetName="ApiSpecPath", Mandatory)]
     [string]$ApiSpecPath,
-    [Parameter(ParameterSetName="ApiSpecPfUrl")]
+    [Parameter(ParameterSetName="ApiSpecPfUrl", Mandatory)]
     [string]$ApiSpecPfUrl,
-    [Parameter(ParameterSetName="ApiSpecPfUrlBranch")]
+    [Parameter(ParameterSetName="ApiSpecPfUrlBranch", Mandatory)]
     [string]$ApiSpecBranch,
     [Parameter(ParameterSetName="ApiSpecGitUrl")]
     [switch]$ApiSpecGitUrl,
+    [string]$OutputPath = "..\..\sdks",
     [switch]$KeepSource,
     [switch]$NonNullable,
     [switch]$Beta
@@ -58,10 +108,15 @@ begin
         "XPlatCppSDK" = "xplatcppsdk";
     }
 
-    $sdksPath = Resolve-Path (Join-Path $PSScriptRoot "..\..\sdks")
+    $sdksPath = $OutputPath
+    if(![System.IO.Path]::IsPathRooted($sdksPath))
+    {
+        $sdksPath = Resolve-Path (Join-Path $PSScriptRoot $OutputPath)
+    }
+
     if(!(Test-Path $sdksPath))
     {
-        mkdir $sdksPath
+        mkdir $sdksPath | Out-Null
     }
 
     if($ApiSpecBranch)
@@ -120,7 +175,7 @@ process
             }
         }
 
-        if(!$KeepSource)
+        if(!$KeepSource -and (Test-Path $destPath))
         {
             Remove-Item $destPath -Recurse -Include *.as, *.cpp, *.cs, *.h, *.java, *.js, *.lua, *.m, *.php, *.py, *.ts
         }
